+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## v.1.0.5
+
+### Features
+
+-- dbt-firebolt now supports the insert/overwrite incremental model.
+
 ## v.1.0.4
 
 ### Features
 
-- dbt-firebolt now supports append-only incremental models.
+- dbt-firebolt now supports the append-only incremental model.
 - We now return better/more accurate responses from the cursor for logging/debugging.
 
 ### Under the hood